exp_implementation/opus4.7_gpt5.5_local_layout_smoke/pyproject.toml
[project]
name = "image-prompt-autoresearch"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "google-genai",
  "torch",
  "transformers",
  "lpips",
  "pillow",
  "numpy",
  "scipy",
  "python-dotenv",
]
